Rename folders and improve project structure
